solving TODOs.
